Update message route and view
This commit updates the message route and view. The path for the message route has been changed to "/m" and the name has been updated to "view-message". The CreateMessageView component has been modified to include the message ID in the message link. The MessageView component has been updated to retrieve the message based on the message ID from the URL.